[FE] [#229]프론트엔드 리팩토링 (#228)
* refector:챗봇 컴포넌트 분류

* refector:로그인 요청 , JWT 디코딩 분리리

* refector:회원가입 요청함수 분리리

* chore:eslint 수정

* refector:카테고리별 영상 모듈화

* chore:eslint 수정

* chore:eslint 수정

* chore: 의존성 배열 수정

* chore 추가 게시판 부분 의존성 배열 오류 해결

* chore: useCallback 삭제

---------
