--- front-end/src/app/watch-histories/page.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Watchhistory = () => {
-  return (
-    <div>
-      <h2>시청기록</h2>
-    </div>
-  );
-};
-
-export default Watchhistory;
